Fix/zios 13735 when pressing cancel on wipe date screen it still wipes data  (#4558)
* do not wipe when confirmText not match

* fix Passcode setup screen is not keyboard avoiding

* guard

+++ Wire-iOS/Sources/UserInterface/Overlay/CustomAppLock/WipeDatabaseViewController.swift
@@ -98,6 +98,10 @@ final class WipeDatabaseViewController: UIViewController {
     func presentConfirmAlert() {
         let confirmController = RequestPasswordController(context: .wiping, callback: { [weak self] confirmText in
             
+            guard confirmText == "wipe_database.alert.confirm_input".localized else {
+                return
+            }
+            
             self?.deleteAccounts()
             Keychain.deletePasscode()
             ///TODO: reset passcode setting to false?
